requirements.txt
fastapi~=0.82.0
tortoise-orm~=0.19.2
asyncpg~=0.26.0
requests~=2.28.1
uvicorn~=0.18.3
pydantic~=1.9.2
typing~=3.7.4